Refactor worklists/schedule CountView's to use one shared view
